Move editor_sectioned_inspector to editor/settings and remove unrelevant includes
